Moves periodic job base class to its own module, like BaseCommand
