+++ bubbles/plugins/__init__.py
@@ -1,8 +1,9 @@
 # import bubbles.plugins.__base__ as base
 # import bubbles.plugins.__base_command__ as cmd
+# import bubbles.plugins.__base_periodic_job__ as job
 
 # so we can just call `from bubbles.plugins import ChatPluginManager`
 # ChatPluginManager = cmd.ChatPluginManager
-# EventLoop = base.EventLoop
+# EventLoop = job.EventLoop
 # BaseCommand = cmd.BaseCommand
-# BasePeriodicJob = base.BasePeriodicJob
+# BasePeriodicJob = job.BasePeriodicJob
